backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.0
pg8000>=1.29.0

# Firebase (keeping for migration period)
firebase-admin==6.2.0
google-cloud-firestore>=2.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator>=2.0.0

# HTTP & API
requests==2.31.0
httpx==0.25.2
twilio==9.1.0

# Scheduling
apscheduler==3.10.4

# Data Processing
pydantic>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Environment & Utils
python-dotenv==1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Performance & Caching
redis>=5.0.0
tenacity>=8.2.3
asyncpg>=0.29.0


